some refacrtoring; better defaults for Blend.fit
